namespace fix and moved address out of trust class
