+++ academiesContracts/Dfe.Academies.Contracts/V4/Trusts/TrustDto.cs
@@ -1,4 +1,4 @@
-﻿namespace Dfe.Academies.Contracts.Trusts;
+﻿namespace Dfe.Academies.Contracts.V4.Trusts;
 
 [Serializable]
 public class TrustDto
@@ -13,20 +13,4 @@ public class TrustDto
 
     public AddressDto Address { get; set; }
 
-}
-
-[Serializable]
-public class AddressDto
-{
-    public string Street { get; set; }
-
-    public string Town { get; set; }
-
-    public string County { get; set; }
-
-    public string Postcode { get; set; }
-
-    public string Locality { get; set; }
-
-    public string Additional { get; set; }
 }
